Move window test cases
