--- modules/siddhi-core/src/test/java/org/wso2/siddhi/core/query/window/external/TimeWindowTestCase.java
@@ -1,197 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.siddhi.core.query.window.external;
-
-import org.apache.log4j.Logger;
-import org.testng.AssertJUnit;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
-import org.wso2.siddhi.core.SiddhiAppRuntime;
-import org.wso2.siddhi.core.SiddhiManager;
-import org.wso2.siddhi.core.event.Event;
-import org.wso2.siddhi.core.query.output.callback.QueryCallback;
-import org.wso2.siddhi.core.stream.input.InputHandler;
-import org.wso2.siddhi.core.stream.output.StreamCallback;
-import org.wso2.siddhi.core.util.EventPrinter;
-
-public class TimeWindowTestCase {
-    private static final Logger log = Logger.getLogger(TimeWindowTestCase.class);
-    private int inEventCount;
-    private int removeEventCount;
-    private boolean eventArrived;
-
-    @BeforeMethod
-    public void init() {
-        inEventCount = 0;
-        removeEventCount = 0;
-        eventArrived = false;
-    }
-
-    @Test
-    public void testTimeWindow1() throws InterruptedException {
-        log.info("TimeWindow Test1");
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-
-        String cseEventStream = "" +
-                "define stream cseEventStream (symbol string, price float, volume int); " +
-                "define window cseEventWindow (symbol string, price float, volume int) time(2 sec) output all events; ";
-
-        String query = "" +
-                "@info(name = 'query0') " +
-                "from cseEventStream " +
-                "insert into cseEventWindow; " +
-                "" +
-                "@info(name = 'query1') " +
-                "from cseEventWindow " +
-                "select symbol,price,volume " +
-                "insert all events into outputStream ;";
-
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
-
-        siddhiAppRuntime.addCallback("query1", new QueryCallback() {
-            @Override
-            public void receive(long timestamp, Event[] inEvents, Event[] removeEvents) {
-                EventPrinter.print(timestamp, inEvents, removeEvents);
-                if (inEvents != null) {
-                    inEventCount = inEventCount + inEvents.length;
-                }
-                if (removeEvents != null) {
-                    AssertJUnit.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
-                    removeEventCount = removeEventCount + removeEvents.length;
-                }
-                eventArrived = true;
-            }
-
-        });
-
-        InputHandler inputHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
-        siddhiAppRuntime.start();
-        inputHandler.send(new Object[]{"IBM", 700f, 0});
-        inputHandler.send(new Object[]{"WSO2", 60.5f, 1});
-        Thread.sleep(4000);
-        AssertJUnit.assertEquals(2, inEventCount);
-        AssertJUnit.assertEquals(2, removeEventCount);
-        AssertJUnit.assertTrue(eventArrived);
-        siddhiAppRuntime.shutdown();
-
-    }
-
-
-    @Test
-    public void testTimeWindow2() throws InterruptedException {
-        log.info("TimeWindow Test2");
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-
-        String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int); " +
-                "define window cseEventWindow (symbol string, price float, volume int) time(1 sec) output all events; ";
-
-        String query = "" +
-                "@info(name = 'query0') " +
-                "from cseEventStream " +
-                "insert into cseEventWindow; " +
-                "" +
-                "@info(name = 'query1') " +
-                "from cseEventWindow " +
-                "select symbol,price,volume " +
-                "insert all events into outputStream ;";
-
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
-
-        siddhiAppRuntime.addCallback("query1", new QueryCallback() {
-            @Override
-            public void receive(long timestamp, Event[] inEvents, Event[] removeEvents) {
-                EventPrinter.print(timestamp, inEvents, removeEvents);
-                if (inEvents != null) {
-                    inEventCount = inEventCount + inEvents.length;
-                }
-                if (removeEvents != null) {
-                    AssertJUnit.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
-                    removeEventCount = removeEventCount + removeEvents.length;
-                }
-                eventArrived = true;
-            }
-
-        });
-
-        InputHandler inputHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
-        siddhiAppRuntime.start();
-        inputHandler.send(new Object[]{"IBM", 700f, 1});
-        inputHandler.send(new Object[]{"WSO2", 60.5f, 2});
-        Thread.sleep(1100);
-        inputHandler.send(new Object[]{"IBM", 700f, 3});
-        inputHandler.send(new Object[]{"WSO2", 60.5f, 4});
-        Thread.sleep(1100);
-        inputHandler.send(new Object[]{"IBM", 700f, 5});
-        inputHandler.send(new Object[]{"WSO2", 60.5f, 6});
-        Thread.sleep(4000);
-        AssertJUnit.assertEquals(6, inEventCount);
-        AssertJUnit.assertEquals(6, removeEventCount);
-        AssertJUnit.assertTrue(eventArrived);
-        siddhiAppRuntime.shutdown();
-
-    }
-
-    @Test
-    public void testTimeWindow3() throws InterruptedException {
-        log.info("TimeWindow Test3");
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-
-
-        String queries = "define stream fireAlarmEventStream (deviceID string, sonar double); " +
-                "define window fireAlarmEventWindow (deviceID string, sonar double) time(30 milliseconds) output " +
-                "expired events; " +
-                "" +
-                "@info(name = 'query0') " +
-                "from fireAlarmEventStream " +
-                "insert into fireAlarmEventWindow; " +
-                "" +
-                "@info(name = 'query1') " +
-                "from fireAlarmEventWindow " +
-                "select deviceID " +
-                "insert expired events into analyzeStream; " +
-                "" +
-                "@info(name = 'query2') " +
-                "from analyzeStream " +
-                "select deviceID " +
-                "insert into bulbOnStream; ";
-
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(queries);
-
-        siddhiAppRuntime.addCallback("analyzeStream", new StreamCallback() {
-            @Override
-            public void receive(Event[] events) {
-                EventPrinter.print(events);
-                eventArrived = true;
-            }
-        });
-
-        InputHandler inputHandler = siddhiAppRuntime.getInputHandler("fireAlarmEventStream");
-        siddhiAppRuntime.start();
-        inputHandler.send(new Object[]{"id1", 20d});
-        inputHandler.send(new Object[]{"id2", 20d});
-        Thread.sleep(2000);
-        AssertJUnit.assertTrue(eventArrived);
-        siddhiAppRuntime.shutdown();
-
-    }
-
-}
