+++ modules/siddhi-core/src/test/java/org/wso2/siddhi/core/window/TimeWindowTestCase.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.siddhi.core.query.window.external;
+package org.wso2.siddhi.core.window;
 
 import org.apache.log4j.Logger;
 import org.testng.AssertJUnit;
@@ -27,10 +27,11 @@ import org.wso2.siddhi.core.SiddhiManager;
 import org.wso2.siddhi.core.event.Event;
 import org.wso2.siddhi.core.query.output.callback.QueryCallback;
 import org.wso2.siddhi.core.stream.input.InputHandler;
+import org.wso2.siddhi.core.stream.output.StreamCallback;
 import org.wso2.siddhi.core.util.EventPrinter;
 
-public class SortWindowTestCase {
-    private static final Logger log = Logger.getLogger(SortWindowTestCase.class);
+public class TimeWindowTestCase {
+    private static final Logger log = Logger.getLogger(TimeWindowTestCase.class);
     private int inEventCount;
     private int removeEventCount;
     private boolean eventArrived;
@@ -43,14 +44,15 @@ public class SortWindowTestCase {
     }
 
     @Test
-    public void testSortWindow1() throws InterruptedException {
-        log.info("SortWindow test1");
+    public void testTimeWindow1() throws InterruptedException {
+        log.info("TimeWindow Test1");
 
         SiddhiManager siddhiManager = new SiddhiManager();
 
         String cseEventStream = "" +
-                "define stream cseEventStream (symbol string, price float, volume long); " +
-                "define window cseEventWindow (symbol string, price float, volume long) sort(2,volume, 'asc'); ";
+                "define stream cseEventStream (symbol string, price float, volume int); " +
+                "define window cseEventWindow (symbol string, price float, volume int) time(2 sec) output all events; ";
+
         String query = "" +
                 "@info(name = 'query0') " +
                 "from cseEventStream " +
@@ -58,7 +60,7 @@ public class SortWindowTestCase {
                 "" +
                 "@info(name = 'query1') " +
                 "from cseEventWindow " +
-                "select volume " +
+                "select symbol,price,volume " +
                 "insert all events into outputStream ;";
 
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
@@ -71,6 +73,7 @@ public class SortWindowTestCase {
                     inEventCount = inEventCount + inEvents.length;
                 }
                 if (removeEvents != null) {
+                    AssertJUnit.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
                     removeEventCount = removeEventCount + removeEvents.length;
                 }
                 eventArrived = true;
@@ -78,72 +81,117 @@ public class SortWindowTestCase {
 
         });
 
+        InputHandler inputHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
+        siddhiAppRuntime.start();
+        inputHandler.send(new Object[]{"IBM", 700f, 0});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 1});
+        Thread.sleep(4000);
+        AssertJUnit.assertEquals(2, inEventCount);
+        AssertJUnit.assertEquals(2, removeEventCount);
+        AssertJUnit.assertTrue(eventArrived);
+        siddhiAppRuntime.shutdown();
+
+    }
+
+
+    @Test
+    public void testTimeWindow2() throws InterruptedException {
+        log.info("TimeWindow Test2");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String cseEventStream = "define stream cseEventStream (symbol string, price float, volume int); " +
+                "define window cseEventWindow (symbol string, price float, volume int) time(1 sec) output all events; ";
+
+        String query = "" +
+                "@info(name = 'query0') " +
+                "from cseEventStream " +
+                "insert into cseEventWindow; " +
+                "" +
+                "@info(name = 'query1') " +
+                "from cseEventWindow " +
+                "select symbol,price,volume " +
+                "insert all events into outputStream ;";
+
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
+
+        siddhiAppRuntime.addCallback("query1", new QueryCallback() {
+            @Override
+            public void receive(long timestamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timestamp, inEvents, removeEvents);
+                if (inEvents != null) {
+                    inEventCount = inEventCount + inEvents.length;
+                }
+                if (removeEvents != null) {
+                    AssertJUnit.assertTrue("InEvents arrived before RemoveEvents", inEventCount > removeEventCount);
+                    removeEventCount = removeEventCount + removeEvents.length;
+                }
+                eventArrived = true;
+            }
+
+        });
 
         InputHandler inputHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
         siddhiAppRuntime.start();
-        inputHandler.send(new Object[]{"WSO2", 55.6f, 100L});
-        inputHandler.send(new Object[]{"IBM", 75.6f, 300L});
-        inputHandler.send(new Object[]{"WSO2", 57.6f, 200L});
-        inputHandler.send(new Object[]{"WSO2", 55.6f, 20L});
-        inputHandler.send(new Object[]{"WSO2", 57.6f, 40L});
-        Thread.sleep(1000);
-        AssertJUnit.assertEquals(5, inEventCount);
-        AssertJUnit.assertEquals(3, removeEventCount);
+        inputHandler.send(new Object[]{"IBM", 700f, 1});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 2});
+        Thread.sleep(1100);
+        inputHandler.send(new Object[]{"IBM", 700f, 3});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 4});
+        Thread.sleep(1100);
+        inputHandler.send(new Object[]{"IBM", 700f, 5});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 6});
+        Thread.sleep(4000);
+        AssertJUnit.assertEquals(6, inEventCount);
+        AssertJUnit.assertEquals(6, removeEventCount);
         AssertJUnit.assertTrue(eventArrived);
         siddhiAppRuntime.shutdown();
 
     }
 
     @Test
-    public void testSortWindow2() throws InterruptedException {
-        log.info("SortWindow test2");
+    public void testTimeWindow3() throws InterruptedException {
+        log.info("TimeWindow Test3");
 
         SiddhiManager siddhiManager = new SiddhiManager();
 
-        String cseEventStream = "" +
-                "define stream cseEventStream (symbol string, price float, volume long); " +
-                "define window cseEventWindow (symbol string, price float, volume long) sort(2,volume, 'asc', price, " +
-                "'desc'); ";
-        String query = "" +
+
+        String queries = "define stream fireAlarmEventStream (deviceID string, sonar double); " +
+                "define window fireAlarmEventWindow (deviceID string, sonar double) time(30 milliseconds) output " +
+                "expired events; " +
+                "" +
                 "@info(name = 'query0') " +
-                "from cseEventStream " +
-                "insert into cseEventWindow; " +
+                "from fireAlarmEventStream " +
+                "insert into fireAlarmEventWindow; " +
                 "" +
                 "@info(name = 'query1') " +
-                "from cseEventWindow " +
-                "select volume " +
-                "insert all events into outputStream ;";
+                "from fireAlarmEventWindow " +
+                "select deviceID " +
+                "insert expired events into analyzeStream; " +
+                "" +
+                "@info(name = 'query2') " +
+                "from analyzeStream " +
+                "select deviceID " +
+                "insert into bulbOnStream; ";
 
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(queries);
 
-        siddhiAppRuntime.addCallback("query1", new QueryCallback() {
+        siddhiAppRuntime.addCallback("analyzeStream", new StreamCallback() {
             @Override
-            public void receive(long timestamp, Event[] inEvents, Event[] removeEvents) {
-                EventPrinter.print(timestamp, inEvents, removeEvents);
-                if (inEvents != null) {
-                    inEventCount = inEventCount + inEvents.length;
-                }
-                if (removeEvents != null) {
-                    removeEventCount = removeEventCount + removeEvents.length;
-                }
+            public void receive(Event[] events) {
+                EventPrinter.print(events);
                 eventArrived = true;
             }
-
         });
 
-
-        InputHandler inputHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
+        InputHandler inputHandler = siddhiAppRuntime.getInputHandler("fireAlarmEventStream");
         siddhiAppRuntime.start();
-        inputHandler.send(new Object[]{"WSO2", 50, 100L});
-        inputHandler.send(new Object[]{"IBM", 20, 100L});
-        inputHandler.send(new Object[]{"WSO2", 40, 50L});
-        inputHandler.send(new Object[]{"WSO2", 100, 20L});
-        inputHandler.send(new Object[]{"WSO2", 50, 50L});
-        Thread.sleep(1000);
-        AssertJUnit.assertEquals(5, inEventCount);
-        AssertJUnit.assertEquals(3, removeEventCount);
+        inputHandler.send(new Object[]{"id1", 20d});
+        inputHandler.send(new Object[]{"id2", 20d});
+        Thread.sleep(2000);
         AssertJUnit.assertTrue(eventArrived);
         siddhiAppRuntime.shutdown();
 
     }
+
 }
