+++ modules/siddhi-core/src/test/resources/testng.xml
@@ -111,17 +111,17 @@
             <class name="org.wso2.siddhi.core.query.window.TimeBatchWindowTestCase"/>
             <class name="org.wso2.siddhi.core.query.window.TimeLengthWindowTestCase"/>
             <class name="org.wso2.siddhi.core.query.window.TimeWindowTestCase"/>
-            <class name="org.wso2.siddhi.core.query.window.external.CronWindowTestCase"/>
-            <class name="org.wso2.siddhi.core.query.window.external.CustomJoinWindowTestCase"/>
-            <class name="org.wso2.siddhi.core.query.window.external.ExternalTimeBatchWindowTestCase"/>
-            <class name="org.wso2.siddhi.core.query.window.external.FrequentWindowEventTbaleTestCase"/>
-            <class name="org.wso2.siddhi.core.query.window.external.LenghtBatchWindowTestCase"/>
-            <class name="org.wso2.siddhi.core.query.window.external.LenghtWindowTestCase"/>
-            <class name="org.wso2.siddhi.core.query.window.external.LossyFrequentWindowTestCase"/>
-            <class name="org.wso2.siddhi.core.query.window.external.SortWindowTestCase"/>
-            <class name="org.wso2.siddhi.core.query.window.external.TimeBatchWindowTestCase"/>
-            <class name="org.wso2.siddhi.core.query.window.external.TimeLengthWindowTestCase"/>
-            <class name="org.wso2.siddhi.core.query.window.external.WindowDefinitionTestCase"/>
+            <class name="org.wso2.siddhi.core.window.CronWindowTestCase"/>
+            <class name="org.wso2.siddhi.core.window.CustomJoinWindowTestCase"/>
+            <class name="org.wso2.siddhi.core.window.ExternalTimeBatchWindowTestCase"/>
+            <class name="org.wso2.siddhi.core.window.FrequentWindowEventTbaleTestCase"/>
+            <class name="org.wso2.siddhi.core.window.LenghtBatchWindowTestCase"/>
+            <class name="org.wso2.siddhi.core.window.LenghtWindowTestCase"/>
+            <class name="org.wso2.siddhi.core.window.LossyFrequentWindowTestCase"/>
+            <class name="org.wso2.siddhi.core.window.SortWindowTestCase"/>
+            <class name="org.wso2.siddhi.core.window.TimeBatchWindowTestCase"/>
+            <class name="org.wso2.siddhi.core.window.TimeLengthWindowTestCase"/>
+            <class name="org.wso2.siddhi.core.window.WindowDefinitionTestCase"/>
 
             <class name="org.wso2.siddhi.core.query.BooleanCompareTestCase"/>
             <class name="org.wso2.siddhi.core.query.FilterTestCase1"/>
